refactor(sandboxes): move the thread sandbox lifecycle out of server.py
Creation, reconnection, proxy-credential refresh, git identity and the
reset/recreate rebinds move from agent/server.py into
agent/sandboxes/lifecycle.py, beside the registry they mutate.

agent/runtime/sandbox.py existed only to let chat/reviewer/analyzer reach
these through function-body imports without closing a cycle back into
server.py; with the lifecycle in its own module agent/runtime re-exports it
directly and the shim is gone.

Test doubles follow the code: patch targets that named agent.server now name
agent.sandboxes.lifecycle. Nothing is re-exported from server.py, so a target
this missed raises AttributeError at patch time rather than silently binding a
name the implementation no longer reads.
