+++ agent/runtime/__init__.py
@@ -1,3 +1,9 @@
+from agent.sandboxes.lifecycle import (
+    configure_git_identity,
+    ensure_sandbox_for_thread,
+    get_cached_sandbox_backend,
+)
+
 from .constants import (
     DEFAULT_LLM_MAX_TOKENS,
     DEFAULT_LLM_MODEL_ID,
@@ -5,11 +11,6 @@ from .constants import (
     MODEL_CALL_RECURSION_LIMIT,
 )
 from .execution import graph_loaded_for_execution
-from .sandbox import (
-    configure_git_identity,
-    ensure_sandbox_for_thread,
-    get_cached_sandbox_backend,
-)
 
 __all__ = [
     "DEFAULT_LLM_MAX_TOKENS",
