+++ tests/sandbox/test_proxy_auth.py
@@ -413,17 +413,21 @@ class TestCreateSandboxWithProxy:
         """Installation token should be used for proxy auth on langsmith sandboxes."""
         with (
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=("ghs_install", None),
             ) as mock_get_token,
-            patch("agent.server.create_sandbox", new_callable=AsyncMock) as mock_create,
-            patch("agent.server._configure_github_proxy", new_callable=AsyncMock) as mock_proxy,
+            patch(
+                "agent.sandboxes.lifecycle.create_sandbox", new_callable=AsyncMock
+            ) as mock_create,
+            patch(
+                "agent.sandboxes.lifecycle._configure_github_proxy", new_callable=AsyncMock
+            ) as mock_proxy,
             patch.dict("os.environ", {"SANDBOX_TYPE": "langsmith", "LANGSMITH_API_KEY": "ls-key"}),
         ):
             mock_create.return_value = MagicMock(id="sandbox-123", aexecute=AsyncMock())
 
-            from agent.server import _create_sandbox_with_proxy
+            from agent.sandboxes.lifecycle import _create_sandbox_with_proxy
 
             await _create_sandbox_with_proxy()
 
@@ -447,22 +451,26 @@ class TestCreateSandboxWithProxy:
         )
         with (
             patch(
-                "agent.server.resolve_environment", new_callable=AsyncMock, return_value=environment
+                "agent.sandboxes.lifecycle.resolve_environment",
+                new_callable=AsyncMock,
+                return_value=environment,
             ),
-            patch("agent.server.create_sandbox", new_callable=AsyncMock) as mock_create,
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.create_sandbox", new_callable=AsyncMock
+            ) as mock_create,
+            patch(
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=("ghs_install", None),
             ),
             patch(
-                "agent.server._configure_github_proxy", new_callable=AsyncMock
+                "agent.sandboxes.lifecycle._configure_github_proxy", new_callable=AsyncMock
             ) as mock_configure_proxy,
             patch.dict("os.environ", {"SANDBOX_TYPE": "langsmith"}),
         ):
             mock_create.return_value = MagicMock(id="sandbox-123", aexecute=AsyncMock())
 
-            from agent.server import _create_sandbox_with_proxy
+            from agent.sandboxes.lifecycle import _create_sandbox_with_proxy
 
             await _create_sandbox_with_proxy(environment_slug="large")
 
@@ -484,17 +492,19 @@ class TestCreateSandboxWithProxy:
         """A full installation token mint failure prevents proxy configuration."""
         with (
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=(None, None),
             ) as mock_get_token,
-            patch("agent.server.create_sandbox", new_callable=AsyncMock) as mock_create,
-            patch("agent.server._configure_github_proxy", new_callable=AsyncMock),
+            patch(
+                "agent.sandboxes.lifecycle.create_sandbox", new_callable=AsyncMock
+            ) as mock_create,
+            patch("agent.sandboxes.lifecycle._configure_github_proxy", new_callable=AsyncMock),
             patch.dict("os.environ", {"SANDBOX_TYPE": "langsmith", "LANGSMITH_API_KEY": "ls-key"}),
         ):
             mock_create.return_value = MagicMock(id="sandbox-123", aexecute=AsyncMock())
 
-            from agent.server import _create_sandbox_with_proxy
+            from agent.sandboxes.lifecycle import _create_sandbox_with_proxy
 
             with pytest.raises(ValueError, match="installation token is unavailable"):
                 await _create_sandbox_with_proxy(thread_id="thread-123")
@@ -505,13 +515,17 @@ class TestCreateSandboxWithProxy:
     async def test_skips_proxy_for_non_langsmith(self) -> None:
         """Non-langsmith sandboxes should skip proxy configuration."""
         with (
-            patch("agent.server.create_sandbox", new_callable=AsyncMock) as mock_create,
-            patch("agent.server._configure_github_proxy", new_callable=AsyncMock) as mock_proxy,
+            patch(
+                "agent.sandboxes.lifecycle.create_sandbox", new_callable=AsyncMock
+            ) as mock_create,
+            patch(
+                "agent.sandboxes.lifecycle._configure_github_proxy", new_callable=AsyncMock
+            ) as mock_proxy,
             patch.dict("os.environ", {"SANDBOX_TYPE": "daytona"}),
         ):
             mock_create.return_value = MagicMock(id="sandbox-456", aexecute=AsyncMock())
 
-            from agent.server import _create_sandbox_with_proxy
+            from agent.sandboxes.lifecycle import _create_sandbox_with_proxy
 
             await _create_sandbox_with_proxy()
 
@@ -522,9 +536,11 @@ class TestCreateSandboxWithProxy:
     async def test_raises_when_no_installation_token_for_langsmith(self) -> None:
         """Should raise ValueError when installation token is unavailable for langsmith."""
         with (
-            patch("agent.server.create_sandbox", new_callable=AsyncMock) as mock_create,
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.create_sandbox", new_callable=AsyncMock
+            ) as mock_create,
+            patch(
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=(None, None),
             ),
@@ -532,7 +548,7 @@ class TestCreateSandboxWithProxy:
         ):
             mock_create.return_value = MagicMock(id="sandbox-789", aexecute=AsyncMock())
 
-            from agent.server import _create_sandbox_with_proxy
+            from agent.sandboxes.lifecycle import _create_sandbox_with_proxy
 
             with pytest.raises(ValueError, match="installation token is unavailable"):
                 await _create_sandbox_with_proxy()
@@ -585,12 +601,12 @@ class TestRefreshProxyOnSandboxReuse:
                 return_value=Environment(slug="env", create_params={"proxy_config": {}}),
             ),
             patch(
-                "agent.server.get_sandbox_id_from_metadata",
+                "agent.sandboxes.lifecycle.get_sandbox_id_from_metadata",
                 new_callable=AsyncMock,
                 return_value="sandbox-cached",
             ),
             patch(
-                "agent.server.get_sandbox_metadata",
+                "agent.sandboxes.lifecycle.get_sandbox_metadata",
                 new_callable=AsyncMock,
                 return_value={
                     "sandbox_id": "sandbox-cached",
@@ -598,11 +614,13 @@ class TestRefreshProxyOnSandboxReuse:
                 },
             ),
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=("ghs_fresh", None),
             ),
-            patch("agent.server._configure_github_proxy", new_callable=AsyncMock) as mock_proxy,
+            patch(
+                "agent.sandboxes.lifecycle._configure_github_proxy", new_callable=AsyncMock
+            ) as mock_proxy,
             patch(
                 "agent.server.aresolve_sandbox_work_dir",
                 new_callable=AsyncMock,
@@ -612,7 +630,7 @@ class TestRefreshProxyOnSandboxReuse:
             patch("agent.server.construct_system_prompt", return_value="prompt"),
             patch("agent.server.create_deep_agent", side_effect=fake_create_deep_agent),
             patch.dict(
-                "agent.server.SANDBOX_BACKENDS",
+                "agent.sandboxes.lifecycle.SANDBOX_BACKENDS",
                 {"thread-123": SandboxBackendProxy(mock_sandbox, thread_id="thread-123")},
                 clear=True,
             ),
@@ -652,26 +670,28 @@ class TestRefreshProxyOnSandboxReuse:
                 return_value=("ghp", None),
             ),
             patch(
-                "agent.server.get_sandbox_id_from_metadata",
+                "agent.sandboxes.lifecycle.get_sandbox_id_from_metadata",
                 new_callable=AsyncMock,
                 return_value="sandbox-existing",
             ),
             patch(
-                "agent.server.get_sandbox_metadata",
+                "agent.sandboxes.lifecycle.get_sandbox_metadata",
                 new_callable=AsyncMock,
                 return_value={"sandbox_id": "sandbox-existing"},
             ),
             patch(
-                "agent.server.create_sandbox",
+                "agent.sandboxes.lifecycle.create_sandbox",
                 new_callable=AsyncMock,
                 return_value=mock_sandbox,
             ) as mock_create,
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=("ghs_fresh", None),
             ),
-            patch("agent.server._configure_github_proxy", new_callable=AsyncMock) as mock_proxy,
+            patch(
+                "agent.sandboxes.lifecycle._configure_github_proxy", new_callable=AsyncMock
+            ) as mock_proxy,
             patch(
                 "agent.server.aresolve_sandbox_work_dir",
                 new_callable=AsyncMock,
@@ -680,7 +700,7 @@ class TestRefreshProxyOnSandboxReuse:
             patch("agent.server.make_model", return_value=MagicMock()),
             patch("agent.server.construct_system_prompt", return_value="prompt"),
             patch("agent.server.create_deep_agent", side_effect=fake_create_deep_agent),
-            patch.dict("agent.server.SANDBOX_BACKENDS", {}, clear=True),
+            patch.dict("agent.sandboxes.lifecycle.SANDBOX_BACKENDS", {}, clear=True),
             patch.dict("agent.utils.github_proxy._PROXY_BASE_CONFIGS", {}, clear=True),
             patch.dict("os.environ", {"SANDBOX_TYPE": "langsmith"}),
         ):
@@ -711,12 +731,12 @@ class TestRefreshProxyOnSandboxReuse:
 
         with (
             patch(
-                "agent.server.get_github_app_installation_token_with_expiry",
+                "agent.sandboxes.lifecycle.get_github_app_installation_token_with_expiry",
                 new_callable=AsyncMock,
                 return_value=("ghs_fresh", None),
             ),
             patch(
-                "agent.server._configure_github_proxy",
+                "agent.sandboxes.lifecycle._configure_github_proxy",
                 new_callable=AsyncMock,
                 side_effect=httpx.HTTPStatusError(
                     "Bad request",
@@ -724,10 +744,15 @@ class TestRefreshProxyOnSandboxReuse:
                     response=response,
                 ),
             ) as mock_proxy,
-            patch("agent.server._create_sandbox_with_proxy", new_callable=AsyncMock) as mock_create,
+            patch(
+                "agent.sandboxes.lifecycle._create_sandbox_with_proxy", new_callable=AsyncMock
+            ) as mock_create,
             patch.dict("os.environ", {"SANDBOX_TYPE": "langsmith"}),
         ):
-            from agent.server import SandboxUnreachableError, _refresh_github_proxy_or_fail
+            from agent.sandboxes.lifecycle import (
+                SandboxUnreachableError,
+                _refresh_github_proxy_or_fail,
+            )
 
             with pytest.raises(SandboxUnreachableError) as excinfo:
                 await _refresh_github_proxy_or_fail(mock_sandbox, "thread-123")
