+++ tests/sandbox/test_sandbox_recovery.py
@@ -66,7 +66,9 @@ async def test_unreachable_sandbox_notifies_then_ends_the_run() -> None:
                 "agent.middleware.tool_error_handler.post_sandbox_unreachable_notification",
                 new_callable=AsyncMock,
             ) as mock_notify,
-            patch("agent.server._create_sandbox_with_proxy", new_callable=AsyncMock) as mock_create,
+            patch(
+                "agent.sandboxes.lifecycle._create_sandbox_with_proxy", new_callable=AsyncMock
+            ) as mock_create,
             pytest.raises(SandboxConnectionError),
         ):
             await middleware.awrap_tool_call(request, handler)
