+++ tests/sandbox/test_sandbox_recreation.py
@@ -2,8 +2,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from agent.sandboxes.lifecycle import recreate_sandbox_for_thread
 from agent.sandboxes.state import SANDBOX_BACKENDS, set_sandbox_backend
-from agent.server import recreate_sandbox_for_thread
 
 
 @pytest.mark.asyncio
@@ -19,18 +19,20 @@ async def test_recreate_sandbox_hands_off_after_metadata_persists() -> None:
 
     with (
         patch(
-            "agent.server.get_sandbox_id_from_metadata",
+            "agent.sandboxes.lifecycle.get_sandbox_id_from_metadata",
             new_callable=AsyncMock,
             return_value="sandbox-old",
         ),
         patch(
-            "agent.server._create_sandbox_with_proxy",
+            "agent.sandboxes.lifecycle._create_sandbox_with_proxy",
             new_callable=AsyncMock,
             return_value=new_sandbox,
         ) as create,
-        patch("agent.server._configure_git_identity", new_callable=AsyncMock) as configure,
         patch(
-            "agent.server.client.threads.update",
+            "agent.sandboxes.lifecycle._configure_git_identity", new_callable=AsyncMock
+        ) as configure,
+        patch(
+            "agent.sandboxes.lifecycle.client.threads.update",
             new_callable=AsyncMock,
             side_effect=persist_metadata,
         ) as update,
@@ -64,18 +66,18 @@ async def test_recreate_sandbox_keeps_old_binding_when_metadata_update_fails() -
 
     with (
         patch(
-            "agent.server.get_sandbox_id_from_metadata",
+            "agent.sandboxes.lifecycle.get_sandbox_id_from_metadata",
             new_callable=AsyncMock,
             return_value="sandbox-old",
         ),
         patch(
-            "agent.server._create_sandbox_with_proxy",
+            "agent.sandboxes.lifecycle._create_sandbox_with_proxy",
             new_callable=AsyncMock,
             return_value=new_sandbox,
         ),
-        patch("agent.server._configure_git_identity", new_callable=AsyncMock),
+        patch("agent.sandboxes.lifecycle._configure_git_identity", new_callable=AsyncMock),
         patch(
-            "agent.server.client.threads.update",
+            "agent.sandboxes.lifecycle.client.threads.update",
             new_callable=AsyncMock,
             side_effect=RuntimeError("metadata unavailable"),
         ),
@@ -97,17 +99,19 @@ async def test_recreate_sandbox_rejects_non_distinct_provider_result() -> None:
 
     with (
         patch(
-            "agent.server.get_sandbox_id_from_metadata",
+            "agent.sandboxes.lifecycle.get_sandbox_id_from_metadata",
             new_callable=AsyncMock,
             return_value="sandbox-same",
         ),
         patch(
-            "agent.server._create_sandbox_with_proxy",
+            "agent.sandboxes.lifecycle._create_sandbox_with_proxy",
             new_callable=AsyncMock,
             return_value=MagicMock(id="sandbox-same"),
         ),
-        patch("agent.server._configure_git_identity", new_callable=AsyncMock) as configure,
-        patch("agent.server.client.threads.update", new_callable=AsyncMock) as update,
+        patch(
+            "agent.sandboxes.lifecycle._configure_git_identity", new_callable=AsyncMock
+        ) as configure,
+        patch("agent.sandboxes.lifecycle.client.threads.update", new_callable=AsyncMock) as update,
     ):
         with pytest.raises(RuntimeError, match="distinct sandbox"):
             await recreate_sandbox_for_thread(thread_id)
